Fixing compiler issues on Windows.

+++ libNeonDomain/include/Neon/domain/details/bGrid/bGrid_imp.h
@@ -307,7 +307,7 @@ auto bGrid<memBlockSizeX, memBlockSizeY, memBlockSizeZ, userBlockSizeX, userBloc
                                                                                *this,
                                                                                lambda,
                                                                                defaultBlockSize,
-                                                                               [](const Neon::index_3d&) { return size_t(0); });
+                                                                               [](const Neon::index_3d&) { return 0; });
     return kContainer;
 }
